+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- The filters pane in course search is now a sliding drawer on mobile.
 - Use SVG icons instead of an icon font.
 
 ### Fixed